Image Compressor
This image compressor uses the singular value decomposition theory in order to acquire a storage-friendly version of any photo. It's features involves compressing a gray-scaled version of the picture, or maintain the RGB scale. This was a project done for the Linear Algebra subject in the university, and will be updated.